moved the get other users from backend-processing to UsersController
