+++ backend-processing.ts
@@ -1,8 +1,3 @@
-interface SimpleUser {
-  id: number;
-  username: string;
-}
-
 export default class BackendProcessing {
   // //returns undefined if fundID is not found
   // retrieveFundName(fundID): Promise<String> {
@@ -12,20 +7,4 @@ export default class BackendProcessing {
   //     });
   //   });
   // }
-
-  getOtherUsers(currentUserID) {
-    return new Promise<SimpleUser[]>((resolve, reject) => {
-      database
-        .getAllUsers()
-        .then((users) => {
-          let filteredList = users.filter((user) => {
-            return user.UserID == currentUserID;
-          });
-          resolve(filteredList);
-        })
-        .catch((error) => {
-          reject(error);
-        });
-    });
-  }
 }
